More lint

+++ app/src/main/java/com/gputreats/orbitalexplorer/MyGL.java
@@ -29,7 +29,7 @@ final class MyGL {
 
     static float[] functionToBuffer2(Function f, double start, double end, int N) {
         float[] data = new float[2 * N];
-        double stepSize = (end - start) / N;
+        double stepSize = (end - start) / (double) N;
         double x = start;
         float value = (float) f.eval(x);
         for (int i = 0; i < N; ++i) {
